requirements.txt
Django~=5.1.2
Pillow~=10.4.0
django-tinymce~=4.1.0